requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
django==4.2.2
django-admin-autocomplete-all==0.6.6
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.2
django-qr-code==3.1.1
django-resized==1.0.2
djangorestframework==3.14.0
drf-nested-routers==0.93.4
exceptiongroup==1.1.1 ; python_version < '3.11'
gunicorn==20.1.0
iniconfig==2.0.0 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
packaging==23.1 ; python_version >= '3.7'
pillow==9.5.0
pluggy==1.0.0 ; python_version >= '3.6'
psycopg2-binary==2.9.6
pydantic==1.10.9 ; python_version >= '3.7'
pypng==0.20220715.0
pytest==7.3.2 ; python_version >= '3.7'
pytest-django==4.5.2
pytz==2023.3
qrcode==7.4.2
segno==1.5.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
setuptools==67.8.0 ; python_version >= '3.7'
sorl-thumbnail==12.9.0
sqlparse==0.4.4 ; python_version >= '3.5'
tomli==2.0.1 ; python_version < '3.11'
typing-extensions==4.6.3 ; python_version >= '3.7'
tzdata==2023.3
werkzeug==2.3.6
